pyproject.toml
[project]
name = "local-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.6.3",
    "flagembedding>=1.3.4",
    "langchain-chroma>=0.2.1",
    "langchain-community>=0.3.17",
    "langchain-openai>=0.3.4",
    "langchain-text-splitters>=0.3.6",
    "langfuse>=2.59.1",
    "pypdf>=5.2.0",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "torch==2.2.0",
]
